requirements.txt
requests
python-dotenv
python-telegram-bot
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
openai
schedule
loguru
aiohttp
python-magic
tenacity
